PullRequest: fastr/2569

+++ com.oracle.truffle.r.native/gnur/patch/src/main/engine.c
@@ -1,6 +1,6 @@
 /*
  *  R : A Computer Language for Statistical Data Analysis
- *  Copyright (C) 2001-2018  The R Core Team.
+ *  Copyright (C) 2001-2020  The R Core Team.
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -1799,7 +1799,7 @@ void GEText(double x, double y, const char * const str, cetype_t enc,
 					wchar_t wc;
 					mbstate_t mb_st;
 					mbs_init(&mb_st);
-					while ((used = mbrtowc(&wc, ss, n, &mb_st)) > 0) {
+					while ((int)(used = mbrtowc(&wc, ss, n, &mb_st)) > 0) {
 #ifdef DEBUG_MI
 					    printf(" centring %s aka %d in MBCS\n", ss, wc);
 #endif
@@ -1817,9 +1817,9 @@ void GEText(double x, double y, const char * const str, cetype_t enc,
 					}
 					done = TRUE;
 				    } else if (enc2 == CE_UTF8) {
-					size_t used;
+					int used;
 					wchar_t wc;
-					while ((used = utf8toucs(&wc, ss)) > 0) {
+					while ((used = (int) utf8toucs(&wc, ss)) > 0) {
 					    if (IS_HIGH_SURROGATE(wc))
 					    	GEMetricInfo(-(int)utf8toucs32(wc, ss), gc, &h, &d, &w, dd);
 					    else
@@ -2644,16 +2644,16 @@ void GEStrMetric(const char *str, cetype_t enc, const pGEcontext gc,
                     wchar_t wc;
                     mbstate_t mb_st;
                     mbs_init(&mb_st);
-                    while ((used = mbrtowc(&wc, s, n, &mb_st)) > 0) {
+                    while ((int)(used = mbrtowc(&wc, s, n, &mb_st)) > 0) {
                         GEMetricInfo((int) wc, gc, &asc, &dsc, &wid, dd);
                         if (asc > *ascent)
                             *ascent = asc;
                         s += used; n -=used;
                     }
                 } else if (enc2 == CE_UTF8) {
-                    size_t used;
+                    int used;
                     wchar_t wc;
-                    while ((used = utf8toucs(&wc, s)) > 0) {
+                    while ((used = (int)utf8toucs(&wc, s)) > 0) {
                     	if (IS_HIGH_SURROGATE(wc))
                     	    GEMetricInfo(-utf8toucs32(wc, s), gc, &asc, &dsc, &wid, dd);
                     	else
@@ -2705,16 +2705,16 @@ void GEStrMetric(const char *str, cetype_t enc, const pGEcontext gc,
                     wchar_t wc;
                     mbstate_t mb_st;
                     mbs_init(&mb_st);
-                    while ((used = mbrtowc(&wc, s, n, &mb_st)) > 0) {
+                    while ((int)(used = mbrtowc(&wc, s, n, &mb_st)) > 0) {
                         GEMetricInfo((int) wc, gc, &asc, &dsc, &wid, dd);
                         if (dsc > *descent)
                             *descent = dsc;
                         s += used; n -=used;
                     }
                 } else if (enc2 == CE_UTF8) {
-                    size_t used;
+                    int used;
                     wchar_t wc;
-                    while ((used = utf8toucs(&wc, s)) > 0) {
+                    while ((used = (int)utf8toucs(&wc, s)) > 0) {
                         if (IS_HIGH_SURROGATE(wc))
                             GEMetricInfo(-utf8toucs32(wc, s), gc, &asc, &dsc, &wid, dd);
                         else
